上传框架代码

+++ base/core.go
@@ -13,7 +13,7 @@ type IClusterService interface {
 	GetTag() string                                                                                                   //获取集群标签
 	GetCategory() core.S_Category                                                                                     //服务类别 例如游戏服
 	GetRpcId() string                                                                                                 //获取rpc通信id
-	GetPreWeight() int32                                                                                              //集群服务负载值 暂时可以不用理会
+	GetPreWeight() float64                                                                                            //集群服务负载值 暂时可以不用理会
 	SetPreWeight(weight int32)                                                                                        //设置服务器权重
 	GetServiceMonitor() core.IServiceMonitor                                                                          //获取监控模块
 	GetSessionsByCategory(category core.S_Category) (ss []core.IServiceSession)                                       //按服务类别获取服务列表
