+++ core/cbase/servicesession.go
@@ -30,17 +30,17 @@ func (this *ServiceSession) GetRpcId() string {
 func (this *ServiceSession) GetType() string {
 	return this.node.Type
 }
-func (this *ServiceSession) GetVersion() int32 {
+func (this *ServiceSession) GetVersion() float32 {
 	return this.node.Version
 }
-func (this *ServiceSession) SetVersion(v int32) {
+func (this *ServiceSession) SetVersion(v float32) {
 	this.node.Version = v
 }
 
-func (this *ServiceSession) GetPreWeight() int32 {
+func (this *ServiceSession) GetPreWeight() float64 {
 	return this.node.PreWeight
 }
-func (this *ServiceSession) SetPreWeight(p int32) {
+func (this *ServiceSession) SetPreWeight(p float64) {
 	this.node.PreWeight = p
 }
 func (this *ServiceSession) Done() {
